+++ paper_uploads/admin/__init__.py
@@ -1,4 +1,4 @@
-from .base import UploadedFileBase  # noqa: F401
+from .base import ResourceAdminBase  # noqa: F401
 from .collection import CollectionAdminBase  # noqa: F401
 from .file import UploadedFileAdmin  # noqa: F401
 from .image import UploadedImageAdmin  # noqa: F401